Move identifier views to own view file out of node
