+++ api/identifiers/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url
 
-from api.nodes import views
+from api.identifiers import views
 
 urlpatterns = [
-    url(r'^(?P<node_identifier>.+)/$', views.NodeIdentifierDetail.as_view(), name=views.NodeIdentifierDetail.view_name),
+    url(r'^(?P<node_identifier>.+)/$', views.IdentifierDetail.as_view(), name=views.IdentifierDetail.view_name),
 ]
