+++ api/identifiers/views.py
@@ -1,0 +1,77 @@
+from modularodm import Q
+from rest_framework import generics, permissions as drf_permissions
+from rest_framework.exceptions import NotFound
+
+
+from framework.auth.oauth_scopes import CoreScopes
+
+from api.base import permissions as base_permissions
+from api.base.views import JSONAPIBaseView
+from api.base.utils import get_object_or_error
+
+from api.identifiers.serializers import IdentifierSerializer
+
+from api.nodes.permissions import (
+    IsPublic,
+    ExcludeWithdrawals,
+)
+
+from website.models import Node
+from website.identifiers.model import Identifier
+
+
+class IdentifierList(JSONAPIBaseView, generics.ListAPIView):
+    """Identifiers for the current node. Read Only for the time being
+    """
+    permission_classes = (
+        IsPublic,
+        drf_permissions.IsAuthenticatedOrReadOnly,
+        base_permissions.TokenHasScope,
+        ExcludeWithdrawals
+    )
+
+    serializer_class = IdentifierSerializer
+
+    view_category = 'nodes'
+    view_name = 'node-identifier-list'
+    node_lookup_url_kwarg = 'node_id'
+
+    def get_node(self, check_object_permissions=True):
+        node = get_object_or_error(
+            Node,
+            self.kwargs[self.node_lookup_url_kwarg],
+            display_name='node'
+        )
+        # Nodes that are folders/collections are treated as a separate resource, so if the client
+        # requests a collection through a node endpoint, we return a 404
+        if node.is_collection:
+            raise NotFound
+        # May raise a permission denied
+        if check_object_permissions:
+            self.check_object_permissions(self.request, node)
+        return node
+
+    # overrides ListCreateAPIView
+    def get_queryset(self):
+        return Identifier.find(Q('referent', 'eq', self.get_node()))
+
+
+class IdentifierDetail(JSONAPIBaseView, generics.RetrieveAPIView):
+    """Identifiers detail for the requested identifier. Read only
+
+    Detail for any identifier attached to a node, including a link back to the node.
+    """
+    permission_classes = (
+        drf_permissions.IsAuthenticatedOrReadOnly,
+    )
+
+    required_read_scopes = [CoreScopes.NODE_CONTRIBUTORS_READ]
+    required_write_scopes = [CoreScopes.NODE_CONTRIBUTORS_WRITE]
+
+    serializer_class = IdentifierSerializer
+    view_category = 'identifiers'
+    view_name = 'node-identifier-detail'
+
+    def get_object(self):
+        identifier = self.kwargs['node_identifier']
+        return Identifier.find_one(Q('value', 'eq', identifier))
